[IMP] kris_project: eliminate installment wizard, reuse installment form
Remove kris.project.installment.wizard and its line model. Move the
pre-populate allocation logic and auto-sequence assignment into
kris.project.installment itself via default_get and create overrides.

Add estimated_amount, allocated_amount, remaining_amount display fields
to kris.project.installment.allocation with a single read_group query
(fixes N+1). Excludes current installment in edit mode to avoid
double-counting (correctness fix).

The installment form is now used for both create and edit, with the
allocation tree showing full info columns identical to the old wizard.
