+++ kris_project/__manifest__.py
@@ -20,7 +20,6 @@
         "views/kris_project_type_views.xml",
         "views/kris_project_allocation_views.xml",
         "views/kris_project_receipt_wizard_views.xml",
-        "views/kris_project_installment_wizard_views.xml",
         "views/kris_project_views.xml",
         "views/kris_project_dashboard_action.xml",
         "views/kris_project_menus.xml",
