+++ kris_project/security/ir.model.access.csv
@@ -30,12 +30,6 @@ access_kris_project_installment_allocation_manager,kris_project_installment_allo
 access_kris_project_receipt_allocation_user,kris_project_receipt_allocation_user,model_kris_project_receipt_allocation,kris_project.group_kris_project_user,1,1,1,1
 access_kris_project_receipt_allocation_officer,kris_project_receipt_allocation_officer,model_kris_project_receipt_allocation,kris_project.group_kris_project_officer,1,1,1,1
 access_kris_project_receipt_allocation_manager,kris_project_receipt_allocation_manager,model_kris_project_receipt_allocation,kris_project.group_kris_project_manager,1,1,1,1
-access_kris_project_installment_wizard_user,kris_project_installment_wizard_user,model_kris_project_installment_wizard,kris_project.group_kris_project_user,1,1,1,1
-access_kris_project_installment_wizard_officer,kris_project_installment_wizard_officer,model_kris_project_installment_wizard,kris_project.group_kris_project_officer,1,1,1,1
-access_kris_project_installment_wizard_manager,kris_project_installment_wizard_manager,model_kris_project_installment_wizard,kris_project.group_kris_project_manager,1,1,1,1
-access_kris_project_installment_wizard_line_user,kris_project_installment_wizard_line_user,model_kris_project_installment_wizard_line,kris_project.group_kris_project_user,1,1,1,1
-access_kris_project_installment_wizard_line_officer,kris_project_installment_wizard_line_officer,model_kris_project_installment_wizard_line,kris_project.group_kris_project_officer,1,1,1,1
-access_kris_project_installment_wizard_line_manager,kris_project_installment_wizard_line_manager,model_kris_project_installment_wizard_line,kris_project.group_kris_project_manager,1,1,1,1
 access_kris_project_receipt_wizard_line_user,kris_project_receipt_wizard_line_user,model_kris_project_receipt_wizard_line,kris_project.group_kris_project_user,1,1,1,1
 access_kris_project_receipt_wizard_line_officer,kris_project_receipt_wizard_line_officer,model_kris_project_receipt_wizard_line,kris_project.group_kris_project_officer,1,1,1,1
 access_kris_project_receipt_wizard_line_manager,kris_project_receipt_wizard_line_manager,model_kris_project_receipt_wizard_line,kris_project.group_kris_project_manager,1,1,1,1
