--- kris_project/views/kris_project_installment_wizard_views.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="view_kris_project_installment_wizard_form" model="ir.ui.view">
-        <field name="name">kris.project.installment.wizard.form</field>
-        <field name="model">kris.project.installment.wizard</field>
-        <field name="arch" type="xml">
-            <form string="เพิ่มงวดงาน">
-                <field name="currency_id" invisible="1"/>
-                <group>
-                    <field name="project_id" readonly="1"/>
-                    <field name="name"/>
-                    <field name="amount"/>
-                    <field name="due_date"/>
-                </group>
-                <separator string="การจัดสรรตามงวด"/>
-                <field name="allocation_ids">
-                    <tree editable="bottom" create="0" delete="0">
-                        <field name="allocation_line_id" invisible="1"/>
-                        <field name="name" readonly="1" string="ผู้รับจัดสรร"/>
-                        <field name="estimated_amount" readonly="1"/>
-                        <field name="allocated_amount" readonly="1"/>
-                        <field name="remaining_amount" readonly="1"/>
-                        <field name="amount"/>
-                        <field name="currency_id" invisible="1"/>
-                    </tree>
-                </field>
-                <footer>
-                    <button
-                        name="action_save"
-                        string="บันทึก"
-                        type="object"
-                        class="btn-primary"
-                    />
-                    <button string="ยกเลิก" class="btn-secondary" special="cancel"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-
-</odoo>
