+++ kris_project/views/kris_project_views.xml
@@ -34,7 +34,11 @@
                 <separator string="การจัดสรรตามงวด"/>
                 <field name="allocation_ids">
                     <tree editable="bottom" create="0" delete="0">
-                        <field name="allocation_line_id" string="ผู้รับจัดสรร" readonly="1"/>
+                        <field name="allocation_line_id" invisible="1"/>
+                        <field name="name" readonly="1" string="ผู้รับจัดสรร"/>
+                        <field name="estimated_amount" readonly="1"/>
+                        <field name="allocated_amount" readonly="1"/>
+                        <field name="remaining_amount" readonly="1"/>
                         <field name="amount" sum="รวม"/>
                         <field name="currency_id" invisible="1"/>
                     </tree>
