+++ kris_project/wizard/__init__.py
@@ -1,2 +1,1 @@
 from . import kris_project_receipt_wizard
-from . import kris_project_installment_wizard
